Split Sanity preview helpers
