+++ sanity/reusableFetch.tsx
@@ -1,5 +1,4 @@
 import { fetchAssetMeta } from "library/sanity/assetMetadata"
-import DraftModeOverlay from "library/sanity/DraftModeOverlay"
 import { draftMode } from "next/headers"
 import type { ClientPerspective, QueryParams } from "next-sanity"
 import { defineLive } from "next-sanity/live"
@@ -11,7 +10,9 @@ import semver from "semver"
 import { Toaster } from "sonner"
 import { FirefoxFix } from "./FirefoxFix"
 import LiveWrapper, { handleError } from "./reusableFetchClient"
+import { SanityPreviewStatusToast } from "./SanityPreviewStatusToast"
 import { SanityLiveProxy } from "./SanityLiveProxy"
+import { SanityVisualEditingOverlay } from "./SanityVisualEditingOverlay"
 
 /**
  * Use defineLive to enable automatic revalidation and refreshing of your fetched content
@@ -92,7 +93,8 @@ export const LibraryLive = async () => {
 		<>
 			{useProxy && <SanityLiveProxy />}
 			<Toaster />
-			{isDraftMode && <DraftModeOverlay />}
+			<SanityPreviewStatusToast isDraftMode={isDraftMode} />
+			<SanityVisualEditingOverlay isDraftMode={isDraftMode} />
 			{isDraftMode && <FirefoxFix />}
 			<LiveWrapper>
 				{(isDraftMode || !useProxy) && (
